Split AugmentJob into AugmentEzbJob and AugmentZdbJob (GOKbIT-139)
