+++ server/gokbg3/grails-app/conf/application.yml
@@ -158,14 +158,14 @@ environments:
                 enabled: false
             packageOaiCaching:
                 enabled: false
-            augment:
-                cron: "0 30 * * * ? *"
             zdbAugment:
                 enabled: false
                 rrCurators: ZDB
+                cron: "0 30 * * * ? *"
             ezbAugment:
                 enabled: false
                 rrCurators: EZB
+                cron: "0 30 2 ? * MON,TUE,WED,THU,FRI *"
             es:
                 cluster: gokbg3-dev
                 host: localhost
@@ -211,14 +211,14 @@ environments:
                 enabled: false
             packageOaiCaching:
                 enabled: false
-            augment:
-                cron: "0 30 * * * ? *"
             zdbAugment:
                 enabled: false
                 rrCurators: ZDB
+                cron: "0 30 * * * ? *"
             ezbAugment:
                 enabled: false
                 rrCurators: EZB
+                cron: "0 30 2 ? * MON,TUE,WED,THU,FRI *"
             es:
                 cluster: gokbg3-test
                 host: localhost
@@ -272,14 +272,14 @@ environments:
                 enabled: false
             packageOaiCaching:
                 enabled: true
-            augment:
-                cron: "0 30 * * * ? *"
             zdbAugment:
                 enabled: false
                 rrCurators: ZDB
+                cron: "0 30 * * * ? *"
             ezbAugment:
                 enabled: false
                 rrCurators: EZB
+                cron: "0 30 2 ? * MON,TUE,WED,THU,FRI *"
             ygorUrl: null
             languagesUrl : null
             es:
